tests/mobile/requirements-mobile.txt
Appium-Python-Client==4.3.0
pytest==8.3.4
PyYAML==6.0.2
pytest-html==4.1.1
pytest-xdist==3.6.1
Pillow==11.1.0
